// src/app/modules/shares/models/respone/dashboard-res.model.ts
export interface DashboardResModel {
  org_unit_code: string;
  org_unit_name: string;
  evaluators_total: number;
  evaluators_count: number;
  unevaluators_count: number;
  score_total: number;
  service_work_total: number;
  service_work_count: number;
  investigative_work_total: number;
  investigative_work_count: number;
  crime_prevention_work_total: number;
  crime_prevention_work_count: number;
  traffic_work_total: number;
  traffic_work_count: number;
  satisfaction_total: number;
  satisfaction_count: number;
  evaluation_date: string;
}